Move barrier helper in support folder
